Add files via upload
This is a file that has an input of two sunspot coordinates, then it creates a 3D-reconstruction of the two sunspots in question. 

Additionally, it calculates the the arc between the two sunspots (not the direct distance, but the geodesic on the surface of the sun - assuming the surface is a smooth sphere).

If you want, you could play around with this and even try it on earth to get a feeling of how the equation works.